requirements.txt
pandas
numpy
seaborn
tqdm
matplotlib
scikit-learn
scipy
dataclasses
catboost
xgboost
lightgbm
dill
ipyleaflet
pystac
pystac_client
odc.stac
planetary_computer
-e .